refactor(lib): relocate IngestStreamEvent contract to @ax/lib/shared
The moved studio consumer can't reach into axctl's ingest dir. Move the
type to @ax/lib/shared/ingest-stream-events; axctl re-exports it (importers
unchanged), studio imports via the @shared alias.
